Change return type to bool for shareOption
Fixes https://github.com/ShekarMudaliyar/social_share/issues/117

The Dart code is expecting a return type of bool and the Android
code is returning a type String.  I'm assuming "success" should
be translated to True.

I looked at the iOS code and it is returning a boolean type for
the same function.

Signed-off-by: Gary O'Neall <gary@sourceauditor.com>